RestaurantDetailContainer 테스트 작성하고 분리&구현
